🔍 refactor.searchBar element 정리
