--- src/elements/SearchBar.tsx
@@ -1,56 +1,0 @@
-import styled from "styled-components";
-import { theme } from "../style/theme";
-import SearchRoundedIcon from "@mui/icons-material/SearchRounded";
-import { useState } from "react";
-import { useNavigate } from "react-router-dom";
-
-export const SearchBar = () => {
-  const [keyword, setKeyword] = useState<string>("");
-  const navigate = useNavigate();
-  const handleSubmit = () => {
-    navigate(`/search/${keyword}`);
-  };
-
-  return (
-    <InputDiv>
-      <SearchInput
-        placeholder="Search"
-        type="text"
-        value={keyword}
-        onChange={(e) => setKeyword(e.target.value)}
-        onKeyDown={(e) => e.key === "Enter" && handleSubmit()}
-      />
-      <SearchIcon />
-    </InputDiv>
-  );
-};
-
-const InputDiv = styled.div`
-  width: 40%;
-  height: 34px;
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-  padding: 0 20px;
-  margin: 4px 20px 0 0;
-  border: 1px solid ${theme.ls03};
-  border-radius: 30px;
-  @media (max-width: 990px) {
-    display: none;
-  }
-`;
-const SearchInput = styled.input`
-  width: 80%;
-  padding: 0 10px 0 0;
-  border: none;
-  :focus {
-    outline: none;
-  }
-  ::placeholder {
-    color: ${theme.fc05};
-    font-size: ${theme.fs14};
-  }
-`;
-const SearchIcon = styled(SearchRoundedIcon)({
-  color: `${theme.fc05}`,
-});
